move all word related logic and routes into a word folder
